+++ src/api/Migrations/20240306155407_RemoveIsLastAttributes.Designer.cs
@@ -3,6 +3,7 @@ using System;
 using BDMS;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore.Infrastructure;
+using Microsoft.EntityFrameworkCore.Migrations;
 using Microsoft.EntityFrameworkCore.Storage.ValueConversion;
 using NetTopologySuite.Geometries;
 using Npgsql.EntityFrameworkCore.PostgreSQL.Metadata;
@@ -12,9 +13,11 @@ using Npgsql.EntityFrameworkCore.PostgreSQL.Metadata;
 namespace BDMS.Migrations
 {
     [DbContext(typeof(BdmsContext))]
-    partial class BdmsContextModelSnapshot : ModelSnapshot
+    [Migration("20240306155407_RemoveIsLastAttributes")]
+    partial class RemoveIsLastAttributes
     {
-        protected override void BuildModel(ModelBuilder modelBuilder)
+        /// <inheritdoc />
+        protected override void BuildTargetModel(ModelBuilder modelBuilder)
         {
 #pragma warning disable 612, 618
             modelBuilder
@@ -54,10 +57,6 @@ namespace BDMS.Migrations
                     b.Property<double?>("FromDepth")
                         .HasColumnType("double precision")
                         .HasColumnName("from_depth");
-
-                    b.Property<bool>("IsOpenBorehole")
-                        .HasColumnType("boolean")
-                        .HasColumnName("is_open_borehole");
 
                     b.Property<int?>("KindId")
                         .HasColumnType("integer")
@@ -940,6 +939,41 @@ namespace BDMS.Migrations
                     b.ToTable("facies_description", "bdms");
                 });
 
+            modelBuilder.Entity("BDMS.Models.Feedback", b =>
+                {
+                    b.Property<int>("Id")
+                        .ValueGeneratedOnAdd()
+                        .HasColumnType("integer")
+                        .HasColumnName("id_feb");
+
+                    NpgsqlPropertyBuilderExtensions.UseIdentityByDefaultColumn(b.Property<int>("Id"));
+
+                    b.Property<DateTime?>("Created")
+                        .HasColumnType("timestamp with time zone")
+                        .HasColumnName("created_feb");
+
+                    b.Property<bool?>("IsFrw")
+                        .HasColumnType("boolean")
+                        .HasColumnName("frw_feb");
+
+                    b.Property<string>("Message")
+                        .HasColumnType("text")
+                        .HasColumnName("message_feb");
+
+                    b.Property<string>("Tag")
+                        .HasColumnType("text")
+                        .HasColumnName("tag_feb");
+
+                    b.Property<string>("User")
+                        .IsRequired()
+                        .HasColumnType("text")
+                        .HasColumnName("user_feb");
+
+                    b.HasKey("Id");
+
+                    b.ToTable("feedbacks", "bdms");
+                });
+
             modelBuilder.Entity("BDMS.Models.File", b =>
                 {
                     b.Property<int>("Id")
@@ -1096,10 +1130,6 @@ namespace BDMS.Migrations
                     b.Property<double?>("FromDepth")
                         .HasColumnType("double precision")
                         .HasColumnName("from_depth");
-
-                    b.Property<bool>("IsOpenBorehole")
-                        .HasColumnType("boolean")
-                        .HasColumnName("is_open_borehole");
 
                     b.Property<int?>("KindId")
                         .HasColumnType("integer")
@@ -2306,7 +2336,7 @@ namespace BDMS.Migrations
                     b.HasOne("BDMS.Models.Codelist", "Kind")
                         .WithMany()
                         .HasForeignKey("KindId")
-                        .OnDelete(DeleteBehavior.NoAction)
+                        .OnDelete(DeleteBehavior.Cascade)
                         .IsRequired();
 
                     b.HasOne("BDMS.Models.User", "UpdatedBy")
@@ -2400,7 +2430,7 @@ namespace BDMS.Migrations
                     b.HasOne("BDMS.Models.Codelist", "Parameter")
                         .WithMany()
                         .HasForeignKey("ParameterId")
-                        .OnDelete(DeleteBehavior.NoAction)
+                        .OnDelete(DeleteBehavior.Cascade)
                         .IsRequired();
 
                     b.HasOne("BDMS.Models.User", "UpdatedBy")
@@ -2767,7 +2797,7 @@ namespace BDMS.Migrations
                     b.HasOne("BDMS.Models.Codelist", "Reliability")
                         .WithMany()
                         .HasForeignKey("ReliabilityId")
-                        .OnDelete(DeleteBehavior.NoAction)
+                        .OnDelete(DeleteBehavior.Cascade)
                         .IsRequired();
 
                     b.HasOne("BDMS.Models.User", "UpdatedBy")
@@ -2862,13 +2892,13 @@ namespace BDMS.Migrations
                     b.HasOne("BDMS.Models.Codelist", "Parameter")
                         .WithMany()
                         .HasForeignKey("ParameterId")
-                        .OnDelete(DeleteBehavior.NoAction)
+                        .OnDelete(DeleteBehavior.Cascade)
                         .IsRequired();
 
                     b.HasOne("BDMS.Models.Codelist", "SampleType")
                         .WithMany()
                         .HasForeignKey("SampleTypeId")
-                        .OnDelete(DeleteBehavior.NoAction)
+                        .OnDelete(DeleteBehavior.Cascade)
                         .IsRequired();
 
                     b.Navigation("Parameter");
@@ -2887,7 +2917,7 @@ namespace BDMS.Migrations
                     b.HasOne("BDMS.Models.Codelist", "Kind")
                         .WithMany()
                         .HasForeignKey("KindId")
-                        .OnDelete(DeleteBehavior.NoAction)
+                        .OnDelete(DeleteBehavior.Cascade)
                         .IsRequired();
 
                     b.Navigation("Kind");
@@ -2917,7 +2947,7 @@ namespace BDMS.Migrations
                     b.HasOne("BDMS.Models.Codelist", "Quantity")
                         .WithMany()
                         .HasForeignKey("QuantityId")
-                        .OnDelete(DeleteBehavior.NoAction)
+                        .OnDelete(DeleteBehavior.Cascade)
                         .IsRequired();
 
                     b.Navigation("Conditions");