+++ src/api/Migrations/20240307094932_FixCodelistDeleteBehavior.Designer.cs
@@ -3,6 +3,7 @@ using System;
 using BDMS;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore.Infrastructure;
+using Microsoft.EntityFrameworkCore.Migrations;
 using Microsoft.EntityFrameworkCore.Storage.ValueConversion;
 using NetTopologySuite.Geometries;
 using Npgsql.EntityFrameworkCore.PostgreSQL.Metadata;
@@ -12,9 +13,11 @@ using Npgsql.EntityFrameworkCore.PostgreSQL.Metadata;
 namespace BDMS.Migrations
 {
     [DbContext(typeof(BdmsContext))]
-    partial class BdmsContextModelSnapshot : ModelSnapshot
+    [Migration("20240307094932_FixCodelistDeleteBehavior")]
+    partial class FixCodelistDeleteBehavior
     {
-        protected override void BuildModel(ModelBuilder modelBuilder)
+        /// <inheritdoc />
+        protected override void BuildTargetModel(ModelBuilder modelBuilder)
         {
 #pragma warning disable 612, 618
             modelBuilder
@@ -54,10 +57,6 @@ namespace BDMS.Migrations
                     b.Property<double?>("FromDepth")
                         .HasColumnType("double precision")
                         .HasColumnName("from_depth");
-
-                    b.Property<bool>("IsOpenBorehole")
-                        .HasColumnType("boolean")
-                        .HasColumnName("is_open_borehole");
 
                     b.Property<int?>("KindId")
                         .HasColumnType("integer")
@@ -620,6 +619,10 @@ namespace BDMS.Migrations
                         .HasColumnType("double precision")
                         .HasColumnName("depth_from");
 
+                    b.Property<bool?>("IsLast")
+                        .HasColumnType("boolean")
+                        .HasColumnName("is_last");
+
                     b.Property<int>("StratigraphyId")
                         .HasColumnType("integer")
                         .HasColumnName("id_sty_fk");
@@ -911,6 +914,10 @@ namespace BDMS.Migrations
                         .HasColumnType("double precision")
                         .HasColumnName("depth_from");
 
+                    b.Property<bool?>("IsLast")
+                        .HasColumnType("boolean")
+                        .HasColumnName("is_last");
+
                     b.Property<int>("StratigraphyId")
                         .HasColumnType("integer")
                         .HasColumnName("id_sty_fk");
@@ -1096,10 +1103,6 @@ namespace BDMS.Migrations
                     b.Property<double?>("FromDepth")
                         .HasColumnType("double precision")
                         .HasColumnName("from_depth");
-
-                    b.Property<bool>("IsOpenBorehole")
-                        .HasColumnType("boolean")
-                        .HasColumnName("is_open_borehole");
 
                     b.Property<int?>("KindId")
                         .HasColumnType("integer")
@@ -1445,6 +1448,10 @@ namespace BDMS.Migrations
                         .HasColumnType("double precision")
                         .HasColumnName("depth_from");
 
+                    b.Property<bool?>("IsLast")
+                        .HasColumnType("boolean")
+                        .HasColumnName("is_last");
+
                     b.Property<int>("StratigraphyId")
                         .HasColumnType("integer")
                         .HasColumnName("id_sty_fk");
@@ -1494,6 +1501,10 @@ namespace BDMS.Migrations
                     b.Property<double?>("FromDepth")
                         .HasColumnType("double precision")
                         .HasColumnName("depth_from");
+
+                    b.Property<bool?>("IsLast")
+                        .HasColumnType("boolean")
+                        .HasColumnName("is_last");
 
                     b.Property<int?>("LithostratigraphyId")
                         .HasColumnType("integer")