+++ CHANGELOG.md
@@ -41,12 +41,14 @@
 - Removed square brackets from _m MD_ in `chronostratigraphy` and `lithostratigraphy`.
 - Implemented a validation to check the user's role and the lock status of the borehole before performing any data manipulations on child tables in the .NET API.
 - Include completion in casing name for hydrogeological observations.
+- Removed property _isLast_ from `ChronostratigraphyLayer`, `LithostratigraphyLayer`, `LithologicalDescription` and `FaciesDescription`.
 - Removed feedback functionality.
 
 ### Fixed
 
 - Ensure hydrogeology components are loaded properly when refreshed.
 - Fix translations for _new_ buttons.
+- Fixed delete behavior for non-nullable foreign keys of type _Codelist_.
 
 ## v2.0.506 - 2023-12-21